BundleDeployment error message improvement
When a resource of a BundleDeployment is removed after the Bundle has
been created, the BundleDeployment says it is missing. This is also the
case if the resource is not removed but changes its owner. In that case,
the message of a resource that is missing might be misleading. These
changes adapt the messsage to say "<resource> is not owned by us"
instead of saying that the "<resource> is missing".

Part of #2134