Prevent `0` as value for gitClientTimeout
A zero value for the gitClientTimeout setting would mean that the
request would not have a client side timeout set, potentially resulting
in an endless waiting state. Although it should not happen that this
value is missing, which would lead to it being zero, it should be safer
to fall back to 30s as in previous versions of Fleet, thereby preventing
the user from disabling the git client timeout completely.

Follow-up for #2188